Remove 404s. Add dedicated list page
