+++ assets/less/helpers.less
@@ -20,3 +20,7 @@
 .ml-3 {
   margin-left: 1rem;
 }
+
+.underlined {
+  text-decoration: underline;
+}
